♻️refactor : prettier 전역 적용

+++ src/services/mutates/post/useHandleLikePost.ts
@@ -1,6 +1,5 @@
 import { useToast } from '@/src/store/useToast'
 import { useMutation } from '@tanstack/react-query'
-
 import { supabase } from '@/src/lib/supabase/client'
 import { getQueryClient } from '@/src/lib/tanstack/get-query-client'
 import { queryKey } from '@/src/lib/tanstack/query-key'
@@ -23,10 +22,7 @@ export default function useHandleLikePost(isLike: boolean | null | undefined) {
       let query: any = supabase.from('like')
 
       if (isLike) {
-        query = query
-          .delete()
-          .eq('user_id', params.meId)
-          .eq('post_id', params.postId)
+        query = query.delete().eq('user_id', params.meId).eq('post_id', params.postId)
       } else {
         query = query
           .insert({
@@ -46,19 +42,10 @@ export default function useHandleLikePost(isLike: boolean | null | undefined) {
     },
     onSettled: (data, error, variables) => {
       openToast({
-        text: isLike
-          ? '좋아하는 게시물에서 삭제하였습니다.'
-          : '좋아하는 게시물로 등록하였습니다.',
+        text: isLike ? '좋아하는 게시물에서 삭제하였습니다.' : '좋아하는 게시물로 등록하였습니다.',
         type: 'info',
       })
-      const {
-        postId,
-        meId,
-        authorId,
-        postType,
-        startOfDay = null,
-        endOfDay = null,
-      } = variables
+      const { postId, meId, authorId, postType, startOfDay = null, endOfDay = null } = variables
 
       void queryClient.invalidateQueries({ queryKey: queryKey.post.public })
       void queryClient.invalidateQueries({
